Simplified language switching
This removes all liveview interactions from the language switching
process.

Since a reload is needed anyways to get the new texts, switching the
language is now reloading the page with a lang param.

This is done by setting the lang attribute on the html element.
An observer then sets the local storage and reloads the page.

fixes #10